Refactor AiDocument and AiParagraph entities to use Long for IDs; enhance AiDocumentQueryDto with knowledgeId and fileName; implement new methods in repositories for batch delete and insert; add uploadFile and vectorDocument methods in AiDocumentService; introduce DocumentSplitterFactory and various splitting strategies; create utility classes for file handling and JSON processing; establish EmbeddingStoreRegistry for managing embedding stores; update application properties for file storage path.

+++ src/main/java/org/hzai/ai/aidocument/entity/AiDocument.java
@@ -30,7 +30,7 @@ public class AiDocument extends PanacheEntityBase {
 	@GeneratedValue
 	private Long id;
 
-	private Integer knowledgeId;
+	private Long knowledgeId;
 
 	private String docName;
